backend/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
httpx
python-dotenv
passlib[bcrypt]
pyjwt
python-multipart
pydantic
